SPAWNING FISH FROM BUTTON

+++ src/global.c
@@ -6,7 +6,7 @@
 
 //struct playerstats pstats = {.money = 0, .fishCount = 0};
 
-playerstats_t create_stats(int64_t money, int fishCount) {
+playerstats_t create_stats(int money, int fishCount) {
 
 	playerstats_t pstats = (playerstats_t){
 		.money = money,
